move mpt witness to zkevm-circuits
