+++ zkevm-circuits/Cargo.toml
@@ -33,6 +33,7 @@ serde.workspace = true
 serde_json.workspace = true
 
 hash-circuit.workspace = true
+mpt-circuits = { package = "halo2-mpt-circuits", git = "https://github.com/scroll-tech/mpt-circuit.git", branch = "v0.7", default-features=false }
 misc-precompiled-circuit = { package = "misc-precompiled-circuit", git = "https://github.com/scroll-tech/misc-precompiled-circuit.git", branch = "main" }
 halo2_gadgets = { git = "https://github.com/scroll-tech/halo2.git", branch = "v1.1", features = ["unstable"] }
 
@@ -64,7 +65,7 @@ warn-unimplemented = ["eth-types/warn-unimplemented"]
 onephase = [] # debug only
 zktrie = []
 poseidon-codehash = []
-parallel_syn = ["hash-circuit/parallel_syn", "halo2_proofs/parallel_syn", "mpt-zktrie/parallel_syn"]
+parallel_syn = ["hash-circuit/parallel_syn", "halo2_proofs/parallel_syn", "mpt-circuits/parallel_syn"]
 
 debug-annotations = []
 enable-stack = ["bus-mapping/enable-stack"]
