+++ zktrie/Cargo.toml
@@ -8,7 +8,6 @@ license.workspace = true
 
 [dependencies]
 halo2_proofs.workspace = true
-mpt-circuits = { package = "halo2-mpt-circuits", git = "https://github.com/scroll-tech/mpt-circuit.git", branch = "v0.7", default-features=false }
 zktrie = { git = "https://github.com/scroll-tech/zktrie.git", branch = "main", features= ["rs_zktrie"] }
 hash-circuit.workspace = true
 eth-types = { path = "../eth-types" }
@@ -23,4 +22,3 @@ serde_json.workspace = true
 
 [features]
 default = []
-parallel_syn = ["mpt-circuits/parallel_syn"]
