+++ zktrie/src/lib.rs
@@ -3,9 +3,11 @@
 #![deny(missing_docs)]
 #![allow(dead_code)]
 
-pub use mpt_circuits;
-// use mpt_circuits::{hash::Hashable, operation::AccountOp, EthTrie,
-// EthTrieCircuit, HashCircuit, MPTProofType};
-
-/// the state modules include structures represent zktrie and witness generator
+/// the state modules include structures represent zktrie and helpers
 pub mod state;
+pub use crate::state::ZktrieState;
+pub use state::builder::{
+    self, extend_address_to_h256, AccountData, AccountProof, BytesArray, CanRead, StorageProof,
+    TrieProof, SECURE_HASH_DOMAIN,
+};
+pub use zktrie::{ZkTrie, ZkTrieNode};
